requirements.txt
google-adk==1.4.2
google-genai
python-dotenv==1.1.0
pillow==11.2.1
litellm
skia-python==138.0
a2a-sdk
uvicorn
nest_asyncio